Ban \p{UnicodeSequenceProperty} at the grammar level
When `CharacterClassEscape` is reached through `ClassEscape`, we know we’re inside of a character class. In that case, only allow the pre-existing `\p{…}` grammar. If not (i.e. when `CharacterClassEscape` is reached through `AtomEscape`), use a new grammar production that matches either the pre-existing `\p{…}` grammar or the new `\p{UnicodeSequenceProperty}` hotness.